Some more review comments
