+++ app/components/edited_indicator/index.tsx
