+++ app/queries/servers/post.ts
@@ -6,6 +6,8 @@ import {of as of$, combineLatestWith} from 'rxjs';
 import {switchMap, distinctUntilChanged} from 'rxjs/operators';
 
 import {MM_TABLES} from '@constants/database';
+import {logDebug, logWarning} from '@utils/log';
+import {updatePermalinkMetadata} from '@utils/permalink_sync';
 
 import {queryGroupsByNames} from './group';
 import {querySavedPostsPreferences} from './preference';
@@ -310,6 +312,60 @@ export const queryPostsWithPermalinkReferences = async (
 
         return referencingPosts;
     } catch (error) {
+        return [];
+    }
+};
+
+/**
+ * Find posts that contain permalink previews referencing the given post ID
+ */
+export const findPostsWithPermalinkReferences = async (
+    database: Database,
+    referencedPostId: string,
+): Promise<PostModel[]> => {
+    try {
+        const referencingPosts = await queryPostsWithPermalinkReferences(database, referencedPostId);
+        return referencingPosts;
+    } catch (error) {
+        logWarning('Error finding posts with permalink references:', error);
+        return [];
+    }
+};
+
+/**
+ * Synchronize permalink previews when a post is edited
+ */
+export const syncPermalinkPreviewsForEditedPost = async (
+    database: Database,
+    editedPost: Post,
+): Promise<PostModel[]> => {
+    try {
+        const referencingPosts = await findPostsWithPermalinkReferences(
+            database,
+            editedPost.id,
+        );
+
+        if (!referencingPosts.length) {
+            return [];
+        }
+
+        const updatedPosts: PostModel[] = [];
+        for (const referencingPost of referencingPosts) {
+            const updatedPost = updatePermalinkMetadata(
+                referencingPost,
+                editedPost.id,
+                editedPost,
+            );
+
+            if (updatedPost) {
+                updatedPosts.push(updatedPost);
+            }
+        }
+
+        logDebug(`Updated ${updatedPosts.length} permalink previews for edited post ${editedPost.id}`);
+        return updatedPosts;
+    } catch (error) {
+        logWarning('Error syncing permalink previews for edited post:', error);
         return [];
     }
 };
