+++ app/utils/permalink_sync.ts
@@ -1,28 +1,9 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import {Database} from '@nozbe/watermelondb';
-
-import {queryPostsWithPermalinkReferences} from '@queries/servers/post';
-import {logDebug, logWarning} from '@utils/log';
+import {logWarning} from '@utils/log';
 
 import type PostModel from '@typings/database/models/servers/post';
-
-/**
- * Find posts that contain permalink previews referencing the given post ID
- */
-export async function findPostsWithPermalinkReferences(
-    database: Database,
-    referencedPostId: string,
-): Promise<PostModel[]> {
-    try {
-        const referencingPosts = await queryPostsWithPermalinkReferences(database, referencedPostId);
-        return referencingPosts;
-    } catch (error) {
-        logWarning('Error finding posts with permalink references:', error);
-        return [];
-    }
-}
 
 /**
  * Update permalink metadata with fresh post data
@@ -82,42 +63,3 @@ export function updatePermalinkMetadata(
         return null;
     }
 }
-
-/**
- * Synchronize permalink previews when a post is edited
- */
-export async function syncPermalinkPreviewsForEditedPost(
-    database: Database,
-    editedPost: Post,
-): Promise<PostModel[]> {
-    try {
-        const referencingPosts = await findPostsWithPermalinkReferences(
-            database,
-            editedPost.id,
-        );
-
-        if (!referencingPosts.length) {
-            return [];
-        }
-
-        const updatedPosts: PostModel[] = [];
-        for (const referencingPost of referencingPosts) {
-            const updatedPost = updatePermalinkMetadata(
-                referencingPost,
-                editedPost.id,
-                editedPost,
-            );
-
-            if (updatedPost) {
-                updatedPosts.push(updatedPost);
-            }
-        }
-
-        logDebug(`Updated ${updatedPosts.length} permalink previews for edited post ${editedPost.id}`);
-        return updatedPosts;
-    } catch (error) {
-        logWarning('Error syncing permalink previews for edited post:', error);
-        return [];
-    }
-}
-
